requirements.txt
streamlit 
Pillow
python-dotenv
openai
langchain
langchain-community
langchain_openai
pinecone